third_party/WebKit/LayoutTests/fast/dom/constructor-proto.html

<!DOCTYPE html>
<script src="../js/resources/js-test-pre.js"></script>
<script>

description('This tests that setting the proto of a built in constructor is not reset');

shouldBe('Object.getPrototypeOf(HTMLElement)', 'Function.prototype');

var proto = new Number(42);
HTMLElement.__proto__ = proto;
shouldBe('Object.getPrototypeOf(HTMLElement)', 'proto');

</script>
<script src="../js/resources/js-test-post.js"></script>
